requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
python-json-logger>=2.0.7
opentelemetry-api>=1.16.0
opentelemetry-sdk>=1.16.0
opentelemetry-instrumentation-fastapi>=0.39b0
# httpx instrumentation only has pre-release/beta builds; allow pre-release installs
opentelemetry-instrumentation-httpx>=0.33b0
opentelemetry-exporter-otlp-proto-grpc>=1.16.0
opentelemetry-exporter-jaeger>=1.16.0

